[Validator] Merge RecursiveValidatorTest with its parents.
